fix : 🚑 reviewDelete  TODO 완료
1. 리뷰 삭제시, 값 변경
(단, 모두 삭제시 기존 값으로 유지 됨._현재데이터들이 리뷰를 다 가져오지 않기 때문에 생기는 문제.)

2. 리뷰 작성은 사용이 완료된 이후 부터 가능  NOT_ALLOW_WRITE_REVIEW(HttpStatus.BAD_REQUEST, "리뷰는 사용이 완료된 후 작성이 가능합니다."),

3. 예약이 삭제되면 roomCnt 값을 올림.

+++ src/main/java/com/zerobase/yogizogi/book/service/BookService.java
@@ -14,6 +14,7 @@ import com.zerobase.yogizogi.user.domain.entity.AppUser;
 import com.zerobase.yogizogi.user.dto.UserDto;
 import com.zerobase.yogizogi.user.repository.UserRepository;
 import com.zerobase.yogizogi.user.token.JwtAuthenticationProvider;
+import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.Objects;
@@ -100,9 +101,8 @@ public class BookService {
         if (provider.validateToken(token)) {
             throw new CustomException(ErrorCode.DO_NOT_ALLOW_TOKEN);
         }
-        // TODO
-        // 예약일이 지난 예약은 삭제 불가능.
-        // 예약 삭제시, (정합하면, Price roomCnt 업데이트)
+
+
         UserDto userDto = provider.getUserDto(token);
         AppUser user = userRepository.findById(userDto.getId())
             .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
@@ -114,10 +114,27 @@ public class BookService {
         Book book = bookRepository.findById(bookId)
             .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_BOOK));
 
+        // 예약일이 지난 예약은 삭제 불가능.
+        LocalDate now = LocalDate.now();
+        if(book.getCheckInDate().isBefore(now)){
+            throw new CustomException(ErrorCode.NOT_ALLOW_DELETE_BOOK);
+        }
+
         if (!Objects.equals(book.getUser().getId(), user.getId())) {
             throw new CustomException(ErrorCode.NOT_ALLOW_ACCESS);
         }
 
+        // 예약 삭제시, (정합하면, Price roomCnt 업데이트)
+        int betweenDay = (int) ChronoUnit.DAYS.between(book.getCheckInDate(), book.getCheckOutDate());
+        IntStream.range(0, betweenDay)
+            .mapToObj(i -> priceRepository.findAllByRoom_IdAndDate(book.getRoom().getId(),
+                book.getCheckInDate().plusDays(i)))
+            .forEach(price -> {price.setRoomCnt(price.getRoomCnt() + 1);
+                priceRepository.save(price);
+            });
+
         bookRepository.delete(book);
+
+
     }
 }
